+++ esp_cam_sensor/sensors/bf3031/include/bf3031.h
@@ -1,0 +1,35 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "esp_cam_sensor_types.h"
+#include "bf3031_types.h"
+
+#if CONFIG_CAMERA_BF3031_SID_HIGH
+#define BF3031_SCCB_ADDR 0x6d
+#else
+#define BF3031_SCCB_ADDR 0x6e
+#endif
+
+#define BF3031_PID         0x3031
+#define BF3031_SENSOR_NAME "BF3031"
+
+/**
+ * @brief Power on camera sensor device and detect the device connected to the designated sccb bus.
+ *
+ * @param[in] config Configuration related to device power-on and detection.
+ * @return
+ *      - Camera device handle on success, otherwise, failed.
+ */
+esp_cam_sensor_device_t *bf3031_detect(esp_cam_sensor_config_t *config);
+
+#ifdef __cplusplus
+}
+#endif