+++ esp_cam_sensor/sensors/bf3031/private_include/bf3031_settings.h
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include <stdio.h>
+#include <stdint.h>
+#include <sdkconfig.h>
+#include "bf3031_regs.h"
+#include "bf3031_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "bf3031_dvp_8bit_20Minput_838x640_yuv422_25fps.h"
+
+#ifdef __cplusplus
+}
+#endif