+++ esp_cam_sensor/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added SC2337P MIPI RAW sensor driver.
+- Added BF3031 DVP Global color sensor.
 
 ## 2.4.0
 - Added SC2331 MIPI RAW sensor driver.